fix: NgIcon cannot be used outside app.component hierarchy/tree; This error was found while trying to use NgIcon in Angular Cdk Dialog where the component lies outside the app html element tree. We instead provideIcons in bootstrapApplication
